@@ -17,7 +17,7 @@ match_memory = "aiutil:memory.main"
 snb = "aiutil.notebook:search.main"
 
 [tool.poetry.dependencies]
-python = ">=3.7.1,<3.11"
+python = ">=3.10,<4.0"
 sqlparse = ">=0.4.1"
 pathspec = ">=0.8.1"
 dateparser = ">=0.7.1"
@@ -46,7 +46,7 @@ PyPDF2 = { version = ">=1.26.0", optional = true }
 # jupyter
 nbformat = { version = ">=5.0.7", optional = true }
 nbconvert = { version = ">=5.6.1", optional = true }
-yapf = { version = ">=0.32.0", optional = true}
+black = "^22.12.0"
 dulwich = ">=0.20.24"
 scikit-image = ">=0.18.3"
 
@@ -54,28 +54,16 @@ scikit-image = ">=0.18.3"
 cv = ["opencv-python", "pillow"]
 docker = ["docker", "networkx", "requests"]
 pdf = ["PyPDF2"]
-jupyter = ["nbformat", "nbconvert", "yapf"]
+jupyter = ["nbformat", "nbconvert", "black"]
 admin = ["psutil"]
-all = ["opencv-python", "pillow", "docker", "networkx", "requests", "PyPDF2", "nbformat", "nbconvert", "yapf", "psutil"]
+all = ["opencv-python", "pillow", "docker", "networkx", "requests", "PyPDF2", "nbformat", "nbconvert", "black", "psutil"]
 
-[tool.poetry.dev-dependencies]
+[tool.poetry.group.dev.dependencies]
 pylint = ">=2.7.0"
 flake8 = ">=3.8.4"
 pytype = { version = ">=2020.08.10", python = "<3.11" }
 darglint = ">=1.5.8"
-yapf = ">=0.32.0"
 pyspark = ">=3.0.1"
-
-[tool.yapf]
-based_on_style = "facebook"
-column_limit = 88
-
-[tool.yapfignore]
-ignore_patterns = [
-    "*/.ipynb_checkpoints/",
-    ".venv/",
-    "venv/",
-]
 
 [tool.pylint.master]
 ignore = ".venv,.ipynb_checkpoints"
