@@ -22,8 +22,8 @@ import aiutil.filesystem as fs
 
 
 class SparkSubmit:
-    """A class for submitting Spark jobs.
-    """
+    """A class for submitting Spark jobs."""
+
     def __init__(self, email: Union[dict, None] = None, level: str = "INFO"):
         """Initialize a SparkSubmit instance.
 
@@ -40,8 +40,11 @@ class SparkSubmit:
         self.email = email
 
     def _spark_log_filter_helper_keyword(
-        self, line: str, keyword: str, mutual_exclusive: list[str],
-        time_delta: datetime.timedelta
+        self,
+        line: str,
+        keyword: str,
+        mutual_exclusive: list[str],
+        time_delta: datetime.timedelta,
     ) -> bool:
         if keyword not in line:
             return False
@@ -58,8 +61,11 @@ class SparkSubmit:
         return False
 
     def _spark_log_filter_helper_keywords(
-        self, line: str, keywords: list[str], mutual_exclusive: bool,
-        time_delta: datetime.timedelta
+        self,
+        line: str,
+        keywords: list[str],
+        mutual_exclusive: bool,
+        time_delta: datetime.timedelta,
     ) -> bool:
         mutual_exclusive = keywords if mutual_exclusive else ()
         for keyword in keywords:
@@ -67,7 +73,7 @@ class SparkSubmit:
                 line=line,
                 keyword=keyword,
                 mutual_exclusive=mutual_exclusive,
-                time_delta=time_delta
+                time_delta=time_delta,
             ):
                 return True
         return False
@@ -78,38 +84,39 @@ class SparkSubmit:
             line=line,
             keywords=["warn client", "uploading"],
             mutual_exclusive=False,
-            time_delta=datetime.timedelta(seconds=0)
+            time_delta=datetime.timedelta(seconds=0),
         ):
             return True
         if self._spark_log_filter_helper_keywords(
             line=line,
             keywords=["queue: ", "tracking url: "],
             mutual_exclusive=False,
-            time_delta=datetime.timedelta(days=1)
+            time_delta=datetime.timedelta(days=1),
         ):
             return True
         if self._spark_log_filter_helper_keywords(
             line=line,
             keywords=["exception", "user class threw", "caused by"],
             mutual_exclusive=False,
-            time_delta=datetime.timedelta(seconds=1)
+            time_delta=datetime.timedelta(seconds=1),
         ):
             return True
         if self._spark_log_filter_helper_keywords(
             line=line,
             keywords=["state: accepted", "state: running", "state: finished"],
             mutual_exclusive=True,
-            time_delta=datetime.timedelta(minutes=10)
+            time_delta=datetime.timedelta(minutes=10),
         ):
             return True
         if self._spark_log_filter_helper_keywords(
             line=line,
             keywords=[
-                "final status: undefined", "final status: succeeded",
-                "final status: failed"
+                "final status: undefined",
+                "final status: succeeded",
+                "final status: failed",
             ],
             mutual_exclusive=True,
-            time_delta=datetime.timedelta(minutes=3)
+            time_delta=datetime.timedelta(minutes=3),
         ):
             return True
         return False
@@ -138,13 +145,17 @@ class SparkSubmit:
         with sp.Popen(cmd, shell=True, stderr=sp.PIPE) as process:
             while True:
                 if process.poll() is None:
-                    line = process.stderr.readline().decode().rstrip()  # pytype: disable=attribute-error
+                    line = (
+                        process.stderr.readline().decode().rstrip()
+                    )  # pytype: disable=attribute-error
                     line = self._filter(line, time_begin, self._spark_log_filter)
                     if line:
                         print(line)
                         stdout.append(line)
                 else:
-                    for line in process.stderr.readlines():  # pytype: disable=attribute-error
+                    for (
+                        line
+                    ) in process.stderr.readlines():  # pytype: disable=attribute-error
                         line = self._filter(
                             line.decode().rstrip(), time_begin, self._spark_log_filter
                         )
@@ -198,7 +209,7 @@ class SparkSubmit:
         lines = fs.filter(
             path=Path(app_id + "_s"),
             pattern=r"^-+\s+Deduped Error Lines\s+-+$",
-            num_lines=999
+            num_lines=999,
         )
         notifiers.get_notifier("email").notify(
             from_=self.email["from"],
@@ -266,8 +277,9 @@ def _get_first_valid_file(key: str, files: list[str]) -> str:
         if _file_exists(file):
             return file
     logger.warning(
-        "None of the specified configuration file for {} exists.\n    ", key,
-        "\n".join("    " + file for file in files)
+        "None of the specified configuration file for {} exists.\n    ",
+        key,
+        "\n".join("    " + file for file in files),
     )
     return ""
 
@@ -311,7 +323,7 @@ def _submit_local(args, config: dict[str, Any]) -> bool:
         raise ValueError(f"{spark_submit} does not exist!")
     lines = [
         "PATH=/usr/local/sbin:/usr/local/bin:/usr/sbin:/usr/bin:/sbin:/bin",
-        spark_submit
+        spark_submit,
     ]
     if config["jars"]:
         lines.append(f"--jars {config['jars']}")
@@ -351,14 +363,17 @@ def _submit_cluster(args, config: dict[str, Any]) -> bool:
         "archives",
         "jars",
     )
-    lines = [config["spark-submit"]] + [
-        f"--{opt} {config[opt]}" for opt in opts if opt in config and config[opt]
-    ] + [f"--conf {k}={v}" for k, v in config["conf"].items()]
+    lines = (
+        [config["spark-submit"]]
+        + [f"--{opt} {config[opt]}" for opt in opts if opt in config and config[opt]]
+        + [f"--conf {k}={v}" for k, v in config["conf"].items()]
+    )
     lines.extend(args.pyfile)
     for idx in range(1, len(lines)):
         lines[idx] = " " * 4 + lines[idx]
-    return SparkSubmit(email=config["email"]
-                      ).submit(" \\\n".join(lines) + "\n", args.pyfile[:1])
+    return SparkSubmit(email=config["email"]).submit(
+        " \\\n".join(lines) + "\n", args.pyfile[:1]
+    )
 
 
 def submit(args: Namespace) -> None:
@@ -401,7 +416,7 @@ def submit(args: Namespace) -> None:
 def parse_args(args=None, namespace=None) -> Namespace:
     """Parse command-line arguments.
 
-    :param args: Arguments to parse. 
+    :param args: Arguments to parse.
     If None, arguments from command line is used.
     :param namespace: An initial Namespace object to use.
     :return: A Namespace object containing command-line options.
@@ -412,21 +427,21 @@ def parse_args(args=None, namespace=None) -> Namespace:
         "--config",
         dest="config",
         default="",
-        help="The configuration file to use."
+        help="The configuration file to use.",
     )
     parser.add_argument(
         "--ssl",
         "--spark-submit-local",
         dest="spark_submit_local",
         default="",
-        help="The local path to spark-submit."
+        help="The local path to spark-submit.",
     )
     parser.add_argument(
         "--pl",
         "--python-local",
         dest="python_local",
         default="",
-        help="The local path to Python."
+        help="The local path to Python.",
     )
     mutex_group = parser.add_mutually_exclusive_group(required=True)
     mutex_group.add_argument(
@@ -434,29 +449,28 @@ def parse_args(args=None, namespace=None) -> Namespace:
         "--gen-config",
         "--generate-config",
         dest="gen_config",
-        help="Specify a path for generating a configration example."
+        help="Specify a path for generating a configration example.",
     )
     mutex_group.add_argument(
         "--py",
         "--pyfile",
         dest="pyfile",
         nargs="+",
-        help="The command (of PySpark script) to submit to Spark to run."
+        help="The command (of PySpark script) to submit to Spark to run.",
     )
     parser.add_argument(
         "--files",
         dest="files",
         nargs="+",
         default=(),
-        help="Additional files to upload."
+        help="Additional files to upload.",
     )
     args = parser.parse_args(args=args, namespace=namespace)
     return args
 
 
 def main():
-    """Define a main function.
-    """
+    """Define a main function."""
     args = parse_args()
     submit(args)
 
